+++ parsedmarc/constants.py
@@ -1,4 +1,4 @@
-__version__ = "11.0.2"
+__version__ = "11.0.3"
 
 USER_AGENT = f"parsedmarc/{__version__}"